26_10_2024_fviz.R
I was looking at a way to remove overlapping 'a' character in the legend ('a' character was overlapping with each point type in the legend). My solution involves the addition of the option show.legend.text=F inside ggscatter to mantain the correct legend, while removing the overlapping character.
I'm new to both R and GitHub, so I hope everything is fine.